- Applied CUDA device handling fixes from commit 5f4422bc7ae4b1c46a4b72be6357ea8c01e9da9e
- Fixed tensor device allocation in test_kullback_leibler.py
- Fixed device type comparisons in test_base.py, test_batchnorm.py, and test_model.py
- Resolved structure conflicts by keeping yetanotherspdnet/ at root

+++ tests/functions/spd_geometries/test_kullback_leibler.py
@@ -338,7 +338,9 @@ class TestArithmeticMean:
             )
         )
         tangent_vectors_last = -torch.sum(tangent_vectors_, dim=0)
-        tangent_vectors = torch.zeros((n_matrices, n_features, n_features), device=device, dtype=dtype)
+        tangent_vectors = torch.zeros(
+            (n_matrices, n_features, n_features), device=device
+        )
         tangent_vectors[:-1] = tangent_vectors_
         tangent_vectors[-1] = tangent_vectors_last
         data = Id + tangent_vectors
@@ -435,7 +437,9 @@ class TestArithmeticMean:
             )
         )
         tangent_vectors_last = -torch.sum(tangent_vectors_, dim=0)
-        tangent_vectors = torch.zeros((n_matrices, n_features, n_features), device=device, dtype=dtype)
+        tangent_vectors = torch.zeros(
+            (n_matrices, n_features, n_features), device=device
+        )
         tangent_vectors[:-1] = tangent_vectors_
         tangent_vectors[-1] = tangent_vectors_last
         # get SPD matrices from tangent vectors
